Refactored Front End
