+++ react-app/src/App.js
@@ -1,11 +1,11 @@
 import React, { useState, useEffect } from "react";
 import { BrowserRouter, Route } from "react-router-dom";
-import LoginForm from "./components/auth/LoginForm";
-import SignUpForm from "./components/auth/SignUpForm";
-import NavBar from "./components/NavBar";
-import ProtectedRoute from "./components/auth/ProtectedRoute";
-import UsersList from "./components/UsersList";
-import User from "./components/User";
+import LoginForm from "./Login/LoginForm";
+import SignUpForm from "./Signup/SignUpForm";
+import ProtectedRoute from "./authorization/ProtectedRoute";
+import UsersList from "./old_components/UsersList";
+import User from "./old_components/User";
+import Homepage from "./Homepage/Homepage";
 import { authenticate } from "./services/auth";
 
 function App() {
@@ -13,7 +13,7 @@ function App() {
   const [loaded, setLoaded] = useState(false);
 
   useEffect(() => {
-    (async() => {
+    (async () => {
       const user = await authenticate();
       if (!user.errors) {
         setAuthenticated(true);
@@ -28,7 +28,6 @@ function App() {
 
   return (
     <BrowserRouter>
-      <NavBar setAuthenticated={setAuthenticated} />
       <Route path="/login" exact={true}>
         <LoginForm
           authenticated={authenticated}
@@ -36,16 +35,23 @@ function App() {
         />
       </Route>
       <Route path="/sign-up" exact={true}>
-        <SignUpForm authenticated={authenticated} setAuthenticated={setAuthenticated} />
+        <SignUpForm
+          authenticated={authenticated}
+          setAuthenticated={setAuthenticated}
+        />
       </Route>
       <ProtectedRoute path="/users" exact={true} authenticated={authenticated}>
-        <UsersList/>
+        <UsersList />
       </ProtectedRoute>
-      <ProtectedRoute path="/users/:userId" exact={true} authenticated={authenticated}>
+      <ProtectedRoute
+        path="/users/:userId"
+        exact={true}
+        authenticated={authenticated}
+      >
         <User />
       </ProtectedRoute>
       <ProtectedRoute path="/" exact={true} authenticated={authenticated}>
-        <h1>My Home Page</h1>
+        <Homepage setAuthenticated={setAuthenticated} />
       </ProtectedRoute>
     </BrowserRouter>
   );
