+++ react-app/src/Homepage/Homepage.js
@@ -1,0 +1,13 @@
+import React from "react";
+import Header from "../shared_components/Header";
+
+const Homepage = ({ setAuthenticated }) => {
+  return (
+    <div>
+      <Header setAuthenticated={setAuthenticated} />
+      <h1>Welcome to Triply!</h1>
+    </div>
+  );
+};
+
+export default Homepage;
