+++ react-app/src/Login/LoginForm.css
@@ -1,0 +1,11 @@
+.login__logo {
+  background-image: url("../assets/triply_logo.png");
+  width: 300px;
+  height: 112px;
+}
+
+.login__container {
+  width: 500px;
+  margin: auto;
+  text-align: -webkit-center;
+}
