+++ react-app/src/shared_components/LogoutButton.js
@@ -1,7 +1,7 @@
 import React from "react";
-import { logout } from "../../services/auth";
+import { logout } from "../services/auth";
 
-const LogoutButton = ({setAuthenticated}) => {
+const LogoutButton = ({ setAuthenticated }) => {
   const onLogout = async (e) => {
     await logout();
     setAuthenticated(false);
