+++ react-app/src/shared_components/NavBar.js
@@ -1,6 +1,6 @@
-import React from 'react';
-import { NavLink } from 'react-router-dom';
-import LogoutButton from './auth/LogoutButton';
+import React from "react";
+import { NavLink } from "react-router-dom";
+import LogoutButton from "./LogoutButton";
 
 const NavBar = ({ setAuthenticated }) => {
   return (
@@ -12,18 +12,13 @@ const NavBar = ({ setAuthenticated }) => {
           </NavLink>
         </li>
         <li>
-          <NavLink to="/login" exact={true} activeClassName="active">
-            Login
-          </NavLink>
-        </li>
-        <li>
           <NavLink to="/sign-up" exact={true} activeClassName="active">
             Sign Up
           </NavLink>
         </li>
         <li>
           <NavLink to="/users" exact={true} activeClassName="active">
-            Users
+            User List
           </NavLink>
         </li>
         <li>
@@ -32,6 +27,6 @@ const NavBar = ({ setAuthenticated }) => {
       </ul>
     </nav>
   );
-}
+};
 
 export default NavBar;
